refactor cuda / c++ kernel code
