+++ cpp/tensorrt_llm/thop/marlinNvfp4MM.cpp
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-#include "tensorrt_llm/kernels/marlin/marlin_nvfp4_gemm.h"
+#include "tensorrt_llm/kernels/marlin/marlin_nvfp4.h"
 #include "tensorrt_llm/thop/thUtils.h"
 #include "userbuffersTensor.h"
 #include <torch/extension.h>
